feat(common): Remove common module (too small)
Features #TT-1
