--- src/common/common.constant.ts
@@ -1,9 +1,0 @@
-// Copyright (c) 2019-present, Anton Makarov <zer0c14@gmail.com>
-// See the LICENSE for more information.
-
-/**
- * Inversion control types.
- */
-export const TYPES = {
-  ConnectionStringParser: Symbol.for('Common\\ConnectionStringParser'),
-};
