--- src/common/index.ts
@@ -1,5 +1,0 @@
-// Copyright (c) 2019-present, Anton Makarov <zer0c14@gmail.com>
-// See the LICENSE for more information.
-
-export { TYPES } from './common.constant';
-export { IConnectionStringParser, parseConnectionString } from './connection.parser';
